+++ src/modules/chat/SearchChannelsWrapper.tsx
@@ -1,29 +1,21 @@
 import ChatPreviewList from '@/components/chats/ChatPreviewList'
 import ChatPreviewSkeleton from '@/components/chats/ChatPreviewSkeleton'
-import ChatSpecialButtons from '@/components/chats/ChatSpecialButtons'
 import NoChatsFound from '@/components/chats/NoChatsFound'
 import useDebounce from '@/hooks/useDebounce'
-import useIsInIframe from '@/hooks/useIsInIframe'
 import useSearch from '@/hooks/useSearch'
 import { getPostsBySpaceContentQuery } from '@/services/subsocial/posts'
 
-export type SearchContentWrapperProps = {
+export type SearchChannelsWrapperProps = {
   children: JSX.Element
   search: string
-  showSpecialButtons?: boolean
-  isIntegrateChatButtonOnTop?: boolean
   getFocusedElementIndex: ReturnType<typeof useSearch>['getFocusedElementIndex']
 }
 
-export default function SearchContentWrapper({
-  isIntegrateChatButtonOnTop,
+export default function SearchChannelsWrapper({
   children,
   search,
-  showSpecialButtons,
   getFocusedElementIndex,
-}: SearchContentWrapperProps) {
-  const isInIframe = useIsInIframe()
-
+}: SearchChannelsWrapperProps) {
   const debouncedSearch = useDebounce(search)
   const { data: searchResults, isLoading } =
     getPostsBySpaceContentQuery.useQuery(search, {
@@ -31,16 +23,7 @@ export default function SearchContentWrapper({
     })
 
   if (!search) {
-    return (
-      <div className='flex flex-col'>
-        {showSpecialButtons && !isInIframe && (
-          <ChatSpecialButtons
-            isIntegrateChatButtonOnTop={!!isIntegrateChatButtonOnTop}
-          />
-        )}
-        {children}
-      </div>
-    )
+    return <div className='flex flex-col'>{children}</div>
   }
 
   if (isLoading) return <ChatPreviewSkeleton.SkeletonList />
